refactor(hooks): hoist generic utility hooks to shared package
