+++ apps/web/src/hooks/useMounted.ts
@@ -1,11 +1,3 @@
 "use client";
 
-import { useSyncExternalStore } from "react";
-
-const subscribe = () => () => undefined;
-const getSnapshot = () => true;
-const getServerSnapshot = () => false;
-
-export function useMounted() {
-  return useSyncExternalStore(subscribe, getSnapshot, getServerSnapshot);
-}
+export { useMounted } from "@esparex/shared";
